Use the `-no-undefined' only for libtool as to not confuse GCC and configure.
In order to convince libtool to create a shared library one need to pass
the `-no-undefined' switch. The most straightforward way is to append it
to LDFLAGS. This approach used to work, however recent GCCs bail out on
any unknown options (sigh!), as a result configure complains that `No
working C compiler was found' and bails out. As a work around introduce
the `EXTRA_LDFLAGS' variable and pass it directly to make.